+++ lib/tomo/testing.rb
@@ -6,14 +6,12 @@ module Tomo
     autoload :CLITester, "tomo/testing/cli_tester"
     autoload :Connection, "tomo/testing/connection"
     autoload :DockerImage, "tomo/testing/docker_image"
-    autoload :DockerPluginTester, "tomo/testing/docker_plugin_tester"
     autoload :HostExtensions, "tomo/testing/host_extensions"
     autoload :Local, "tomo/testing/local"
     autoload :LogCapturing, "tomo/testing/log_capturing"
     autoload :MockedExecError, "tomo/testing/mocked_exec_error"
     autoload :MockedExitError, "tomo/testing/mocked_exit_error"
     autoload :MockPluginTester, "tomo/testing/mock_plugin_tester"
-    autoload :PluginTester, "tomo/testing/plugin_tester"
     autoload :RemoteExtensions, "tomo/testing/remote_extensions"
     autoload :SSHExtensions, "tomo/testing/ssh_extensions"
