+++ lib/tomo/testing/Dockerfile
@@ -3,8 +3,6 @@ WORKDIR /provision
 COPY ./tomo_test_ed25519.pub /root/.ssh/authorized_keys
 COPY ./ubuntu_setup.sh ./
 RUN ./ubuntu_setup.sh
-COPY ./custom_setup.sh ./
-RUN ./custom_setup.sh
 COPY ./systemctl.rb /usr/local/bin/systemctl
 RUN chmod a+x /usr/local/bin/systemctl
 EXPOSE 22
